numa: numa_node_size.c: Fix incompatible pointer type
      CC  numa_node_size64.c to numa_node_size64.o
    numa_node_size64.c: In function ‘main’:
    numa_node_size64.c:10:33: error: passing argument 2 of ‘numa_node_size’ from incompatible pointer type [-Werror=incompatible-pointer-types]
       10 |         ret = numa_node_size(0, &freep);
          |                                 ^~~~~~
          |                                 |
          |                                 long int *
    In file included from numa_node_size64.c:2:
    /usr/include/numa.h:146:47: note: expected ‘long long int *’ but argument is of type ‘long int *’
      146 | long long numa_node_size(int node, long long *freep);
          |                                    ~~~~~~~~~~~^~~~~
    cc1: all warnings being treated as errors
    make: *** [Makefile:47: numa_node_size64.o] Error 1

+++ numa/samples/numa_node_size.c
@@ -4,11 +4,11 @@
 int main(void)
 {
 	long ret;
-	long freep = 0;
+	long long freep = 0;
 	long long freep64 = 0;
 
 	ret = numa_node_size(0, &freep);
-	printf("numa_node_size : ret = %ld G, freep = %ld G\n",
+	printf("numa_node_size : ret = %ld G, freep = %lld G\n",
 		ret / 1024 / 1024, freep / 1024 / 1024);
 
 	ret = numa_node_size64(0, &freep64);
